Refactor stuff

+++ routes/api.php
@@ -44,14 +44,14 @@ Route::prefix('v1')->group(function () {
     });
 
     Route::group(['middleware' => 'custom.jwt.auth'], function () {
+        Route::delete('/users/{id}', 'Api\Auth\AuthController@deleteUser');
+
         Route::group(['namespace' => 'Api\User'], function () {
             Route::get('/notifications', 'UserNotificationsController@getNotifications');
             Route::get('/notifications/unread', 'UserNotificationsController@getUnreadNotifications');
             Route::put('/notifications/read', 'UserNotificationsController@readNotifications');
             Route::delete('/notifications', 'UserNotificationsController@deleteNotifications');
         });
-
-        Route::delete('/users/{id}', 'AuthController@deleteUser');
 
         Route::get('/users/{user_id}/lists', 'Api\User\UserList\UserListController@userLists')
             ->name('user-list.lists');
